correct empty yaml value
Attempting to fix the Travis Error being thrown. Unclear *which* file
is making it barf, other than that it's a line 19. This doc has an
empty value at line 19 and is in the commit that made it barf. perhaps
this is the culprit. (NB build is not failing for me locally)